fix: rename reserved 'pages' directory
