+++ .env.example
@@ -1,0 +1,9 @@
+NEXT_PUBLIC_FIREBASE_API_KEY=YOUR_VALUE
+NEXT_PUBLIC_FIREBASE_AUTH_DOMAIN=YOUR_VALUE
+NEXT_PUBLIC_FIREBASE_PROJECT_ID=YOUR_VALUE
+NEXT_PUBLIC_FIREBASE_STORAGE_BUCKET=YOUR_VALUE
+NEXT_PUBLIC_FIREBASE_MESSAGING_SENDER_ID=YOUR_VALUE
+NEXT_PUBLIC_FIREBASE_APP_ID=YOUR_VALUE
+FIREBASE_PROJECT_ID=YOUR_VALUE
+FIREBASE_CLIENT_EMAIL=YOUR_VALUE
+FIREBASE_PRIVATE_KEY=YOUR_VALUE
