+++ src/app/[locale]/variables/page.tsx
@@ -4,10 +4,13 @@ import dynamic from 'next/dynamic';
 
 import Loader from '@/components/layout/loader/loader';
 
-const Variables = dynamic(() => import('@/pages/variables/variables'), {
-  ssr: false,
-  loading: Loader,
-});
+const Variables = dynamic(
+  () => import('@/pages-components/variables/variables'),
+  {
+    ssr: false,
+    loading: Loader,
+  }
+);
 
 export default function VariablesPage() {
   return <Variables />;
