+++ src/pages-components/main/mainPageClientWrapper.tsx
@@ -4,10 +4,13 @@ import dynamic from 'next/dynamic';
 
 import Loading from '@/components/layout/loader/loader';
 
-const HomePageClient = dynamic(() => import('@/pages/main/mainPageClient'), {
-  ssr: false,
-  loading: () => <Loading />,
-});
+const HomePageClient = dynamic(
+  () => import('@/pages-components/main/mainPageClient'),
+  {
+    ssr: false,
+    loading: () => <Loading />,
+  }
+);
 
 interface HomePageClientWrapperProps {
   currentUser: string;
